requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.18
SQLAlchemy==0.9.2
Tempita==0.5.3dev
Werkzeug==0.9.4
decorator==3.4.0
gunicorn==18.0
itsdangerous==0.23
pbr==0.6
psycopg2==2.5.2
sqlalchemy-migrate==0.8.2
wsgiref==0.1.2
